KAS-1408: uitfaseren print-overview mixin
