+++ app/templates/print-overviews/press-agenda.hbs
@@ -3,7 +3,6 @@
   title=(await title)
   model=(await model)
   routeModelPrefix=routeModel
-  print=(action "print")
   navigateBackToAgenda=(action "navigateBackToAgenda")
 }}
 <div class="vlc-scroll-wrapper__body">
